Adding better code layout.
