+++ lib/xombot/plugins/hello.rb
@@ -1,0 +1,10 @@
+class Hello
+  include Cinch::Plugin
+
+  match "hello"
+  help "This command says hello."
+
+  def execute(m)
+    m.reply "Hello, #{m.user.nick}"
+  end
+end
